+++ src/runtime/values/functions/base_function.py
@@ -11,7 +11,7 @@
 # nougaro modules imports
 from src.lexer.position import Position
 from src.runtime.values.basevalues.value import Value
-from src.runtime.values.basevalues.basevalues import NoneValue, String, Number
+from src.runtime.values.basevalues.basevalues import NoneValue, String, Number, DefaultValue
 from src.runtime.runtime_result import RTResult
 from src.errors.errors import RunTimeError
 from src.runtime.context import Context
@@ -22,7 +22,9 @@ from src.runtime.symbol_table import SymbolTable
 
 class BaseFunction(Value):
     """Parent class for all the function classes (Function, BaseBuiltinFunction and its children)"""
-    def __init__(self, name: str | None, pos_start: Position, pos_end: Position, call_with_module_context: bool = False):
+    def __init__(
+            self, name: str | None, pos_start: Position, pos_end: Position, call_with_module_context: bool = False
+    ):
         super().__init__(pos_start, pos_end)
         self.name: str = name if name is not None else '<function>'
         # (if 'name' is None, we have something like `def()->foo`)
@@ -52,10 +54,11 @@ class BaseFunction(Value):
             new_context.symbol_table = SymbolTable(new_context.parent.symbol_table)
         return new_context
 
-    def check_args(self, param_names: list[str], args: list[Value | tuple[String, Value]], optional_params: list[str] | None = None,
+    def check_args(self, param_names: list[str], args: list[Value],
+                   optional_params: list[tuple[str, Value | None]] | None = None,
                    should_respect_args_number: bool = True):
         # this function is not overwritten
-        """Check if the number of args match with the number of params/optional params"""
+        """Check if the number of args match with the number of params/optional params."""
         # create a result
         result = RTResult()
         if optional_params is None:  # there is no optional params
@@ -65,10 +68,6 @@ class BaseFunction(Value):
         if (len(args) > len(param_names + optional_params)) and should_respect_args_number:
             first_arg = args[len(param_names + optional_params)]
             last_arg = args[-1]
-            if isinstance(first_arg, tuple):
-                first_arg = first_arg[1]
-            if isinstance(last_arg, tuple):
-                last_arg = last_arg[1]
             assert self.context is not None
             return result.failure(RunTimeError(
                 first_arg.pos_start, last_arg.pos_end,
@@ -84,160 +83,68 @@ class BaseFunction(Value):
                 f"{len(param_names) - len(args)} too few args passed into '{self.name}'.",
                 self.context, origin_file="src.values.functions.base_function.BaseFunction.check_args"
             ))
-        
-        # check names
-        found_keyword_arg = False
-        populated_params: list[str] = []
-        populated_optional_params: list[str] = []
-        for i, argument in enumerate(args):
-            if found_keyword_arg and isinstance(argument, Value):
-                assert self.context is not None
-                return result.failure(RunTimeError(
-                    argument.pos_start, argument.pos_end,
-                    "positional arguments should not be placed after keyword arguments.",
-                    self.context, origin_file="src.values.functions.base_funcntion.BaseFunction.check_args"
-                ))
-            if isinstance(argument, tuple):
-                argument_name, argument_value = argument
-                if argument_name.value in param_names:
-                    if argument_name.value in populated_params:
-                        assert self.context is not None
-                        return result.failure(RunTimeError(
-                            argument_name.pos_start, argument_value.pos_end,
-                            f"parameter '{argument_name.value}' is already populated.",
-                            self.context, origin_file="src.values.functions.base_funcntion.BaseFunction.check_args"
-                        ))
-                    populated_params.append(argument_name.value)
-                elif argument_name.value in optional_params:
-                    if argument_name.value in populated_optional_params:
-                        assert self.context is not None
-                        return result.failure(RunTimeError(
-                            argument_name.pos_start, argument_value.pos_end,
-                            f"optional parameter '{argument_name.value}' is already populated.",
-                            self.context, origin_file="src.values.functions.base_funcntion.BaseFunction.check_args"
-                        ))
-                    populated_optional_params.append(argument_name.value)
-            elif i < len(param_names):
-                populated_params.append(param_names[i])
-            elif len(param_names) + len(optional_params) == 0:
-                if not should_respect_args_number:
-                    break
-                # the following code is not gonna execute… Or is it?
-                assert self.context is not None
-                return result.failure(RunTimeError(
-                    argument.pos_start, argument.pos_end,
-                    f"function '{self.name}' takes no argument. If you see this message, "
-                    f"PLEASE OPEN AN ISSUE on https://github.com/jd-develop/nougaro/issues/new/choose with "
-                    f"all the informations.",
-                    self.context, origin_file="src.values.functions.base_funcntion.BaseFunction.check_args"
-                ))
-            elif i-len(param_names) >= len(optional_params):
-                if not should_respect_args_number:
-                    break
-                # the following code is not gonna execute… Or is it?
-                assert self.context is not None
-                return result.failure(RunTimeError(
-                    argument.pos_start, argument.pos_end,
-                    f"too many arguments given to function '{self.name}'. If you see this message, "
-                    f"PLEASE OPEN AN ISSUE on https://github.com/jd-develop/nougaro/issues/new/choose with "
-                    f"all the informations.",
-                    self.context, origin_file="src.values.functions.base_funcntion.BaseFunction.check_args"
-                ))
-            else:
-                populated_optional_params.append(optional_params[i-len(param_names)])
-
-        if sorted(populated_params) != sorted(param_names):
-            first_arg = args[len(param_names + optional_params)]
-            last_arg = args[-1]
-            if isinstance(first_arg, tuple):
-                first_arg = first_arg[1]
-            if isinstance(last_arg, tuple):
-                last_arg = last_arg[1]
-            assert self.context is not None
-            return result.failure(RunTimeError(
-                first_arg.pos_start, last_arg.pos_end,
-                f"too much or too few mendatory arguments were passed into '{self.name}'.",
-                self.context, origin_file="src.values.functions.base_function.BaseFunction.check_args"
-            ))
 
         return result.success(NoneValue(self.pos_start, self.pos_end))  # if there is the right number of params
 
     @staticmethod
-    def populate_args(param_names: list[str], args: list[Value | tuple[String, Value]],
-                      exec_context: Context, optional_params: list[str] | None = None,
+    def populate_args(param_names: list[str], args: list[Value],
+                      exec_context: Context, optional_params: list[tuple[str, Value | None]] | None = None,
                       should_respect_args_number: bool = True):
         # this function is not overwritten
-        """Make the args match to the param names in the symbol table"""
+        """Make the args match to the param names in the symbol table.
+        
+        Any <default> in the args must be replaced with a DefaultValue.
+        """
         # We need the context for the symbol table :)
         if optional_params is None:  # there is no optional params
             optional_params = []
 
         assert exec_context.symbol_table is not None
         if should_respect_args_number:  # the number of args SHOULD be equal to the number of params
-            found_keyword_argument = False
             for i in range(len(args)):
-                if found_keyword_argument:
-                    argument = args[i]
-                    assert isinstance(argument, tuple)  # this has been properly checked in check_args
-                    arg_name = argument[0].value
-                    arg_value = argument[1]
-                    arg_value.set_context(exec_context)
-                    exec_context.symbol_table.set(arg_name, arg_value)
-                elif i < len(param_names):  # the argument is in the non-optional parameters list
-                    arg_name = param_names[i]
+                if i < len(param_names):  # the argument is in the non-optional parameters list
+                    param_name = param_names[i]
                     arg_value = args[i]
-                    if not isinstance(arg_value, Value):
-                        arg_name = arg_value[0].value
-                        arg_value = arg_value[1]
-                        found_keyword_argument = True
                     arg_value.set_context(exec_context)
-                    exec_context.symbol_table.set(arg_name, arg_value)
+                    exec_context.symbol_table.set(param_name, arg_value)
                 else:  # the argument is in the optional parameters list
-                    arg_name = optional_params[len(param_names) - i]
+                    param_name, default_value = optional_params[i - len(param_names)]
                     arg_value = args[i]
-                    if not isinstance(arg_value, Value):
-                        arg_name = arg_value[0].value
-                        arg_value = arg_value[1]
-                        found_keyword_argument = True
                     arg_value.set_context(exec_context)
-                    exec_context.symbol_table.set(arg_name, arg_value)
+                    if isinstance(arg_value, DefaultValue):
+                        arg_value = default_value
+                    if arg_value is not None:
+                        exec_context.symbol_table.set(param_name, arg_value)
+            for i in range(len(param_names) + len(optional_params) - len(args)):
+                # we still have to set those optional args to their default values
+                param_name, default_value = optional_params[len(args) - len(param_names) + i]
+                if default_value is not None:
+                    exec_context.symbol_table.set(param_name, default_value)
         else:  # the number of args may not be equal to the number of params
-            found_keyword_argument = False
             for i in range(len(args)):
-                if found_keyword_argument:
-                    argument = args[i]
-                    assert isinstance(argument, tuple)  # this has been properly checked in check_args
-                    arg_name = argument[0].value
-                    arg_value = argument[1]
-                    arg_value.set_context(exec_context)
-                    exec_context.symbol_table.set(arg_name, arg_value)
-                elif i < len(param_names):  # the argument is in the non-optional parameters list (when this happens ?)
-                    arg_name = param_names[i]
+                if i < len(param_names):  # the argument is in the non-optional parameters list (when this happens ?)
+                    param_name = param_names[i]
                     arg_value = args[i]
-                    if not isinstance(arg_value, Value):
-                        arg_name = arg_value[0].value
-                        arg_value = arg_value[1]
-                        found_keyword_argument = True
                     arg_value.set_context(exec_context)
-                    exec_context.symbol_table.set(arg_name, arg_value)
+                    exec_context.symbol_table.set(param_name, arg_value)
                 elif (i - len(param_names)) < len(optional_params):
                     # the argument is in the optional parameters list
-                    arg_name = optional_params[(i - len(param_names))]
+                    param_name, default_value = optional_params[i - len(param_names)]
                     arg_value = args[i]
-                    if not isinstance(arg_value, Value):
-                        arg_name = arg_value[0].value
-                        arg_value = arg_value[1]
-                        found_keyword_argument = True
                     arg_value.set_context(exec_context)
-                    exec_context.symbol_table.set(arg_name, arg_value)
+                    if isinstance(arg_value, DefaultValue):
+                        arg_value = default_value
+                    if arg_value is not None:
+                        exec_context.symbol_table.set(param_name, arg_value)
                 else:  # the argument is not in the non-optional parameters list, nor in the optional
                     # we don't need to continue to loop: we are at after the last argument that can be matched with
                     # a param. Breaking here may improve performance in some specific contexts
                     break
 
     def check_and_populate_args(
-            self, param_names: list[str], args: list[Value | tuple[String, Value]], exec_context: Context,
-            optional_params: list[str] | None = None, should_respect_args_number: bool = True
+            self, param_names: list[str], args: list[Value], exec_context: Context,
+            optional_params: list[tuple[str, Value | None]] | None = None,
+            should_respect_args_number: bool = True
     ) -> RTResult:
         """self.check_args() then self.populate_args()"""
         # We still need the context for the symbol table